pyproject.toml
[project]
name = "impulse"
version = "0.2.0"
description = "Serverless document processing pipeline on AWS"
readme = "README.md"
requires-python = ">=3.11,<3.14"

dependencies = [
  # Core
  "boto3>=1.42.2",
  "certifi>=2024.0.0",
  "loguru>=0.7.3",
  "pymongo>=4.15.2",

  # Image processing
  "opencv-python>=4.12.0.88",
  "pillow>=10.4.0",

  # Document extraction
  "marker-pdf>=1.10.2",
  "pymupdf>=1.27.1",
  "lxml>=6.0.2",

  # NLP / NER
  "spacy>=3.8.11",
  "transformers>=4.57.0",

  # HTTP (for geocoding)
  "requests>=2.31.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "moto[s3,sqs,ecs]>=5.0.0",
  "mongomock>=4.1.0",
]

[tool.basedpyright]
typeCheckingMode = "basic"
diagnosticMode = "openFilesOnly"
useLibraryCodeForTypes = false
include = ["src"]
executionEnvironments = [{ root = "src" }]
venvPath = "."
venv = ".venv"
exclude = [
  "**/node_modules",
  "**/__pycache__",
  "**/.venv",
  "**/dist",
  "**/build",
  "**/.git",
  "infra/",
  "frontend/",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["impulse*"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
